+++ CHANGELOG.md
@@ -14,6 +14,10 @@ For a full diff see [`0.6.1...main`][0.6.1...main].
 - Added form to enter reading progress directly in the book details modal
 - Added option to sort books by reading progress
 
+### Changed
+
+- Removed 'year' option from pages read bar chart when a year is selected
+
 ## [`0.6.1`][0.6.1] - 2025-03-30
 
 For a full diff see [`0.6.0...0.6.1`][0.6.0...0.6.1].